Creating a I2C communication

+++ Arduino-Control/communication-Rasp/pid.hpp
@@ -3,16 +3,16 @@ template <class T>
 class PID{
 public:
     
-    double error;
+    int error;
     T sample;
     T lastSample;
-    double kP, kI, kD;
-    double I;
+    int kP, kI, kD;
+    int I;
     
     T setPoint;
     long lastProcess;
     
-    PID(double _kP, double _kI, double _kD, T _set){
+    PID(int _kP, int _kI, int _kD, T _set){
         kP = _kP;
         kI = _kI;
         kD = _kD;
@@ -20,7 +20,7 @@ public:
         setSetPoint(_set);
     }
     
-    double addNewSample(T _sample){
+    int addNewSample(T _sample){
         sample = _sample;
         
         return process();
@@ -30,18 +30,18 @@ public:
         setPoint = _setPoint;
     }
     
-    double process(){
+    int process(){
         error = setPoint - sample;
         
-        float deltaTime = (millis() - lastProcess) / 1000.0;
+        int deltaTime = (millis() - lastProcess) / 1000;
         lastProcess = millis();
         
-        double P = error * kP;
+        int P = error * kP;
         I = I + (error * kI) * deltaTime;
-        double D = (lastSample - sample) * kD / deltaTime;
+        int D = (lastSample - sample) * kD / deltaTime;
         lastSample = sample;
         
-        double pid = P + I + D;
+        int pid = P + I + D;
         
         return pid;
     }
