+++ Arduino-Control/communication-Rasp/vector.hpp
@@ -3,23 +3,23 @@
 // Structure to represent vector and vector operations
 // Operations based on angles
 struct vector{
-    float xPos;
-    float yPos;
+    int xPos;
+    int yPos;
     
-    vector(float x = 0.0, float y = 0.0)
+    vector(int x = 0, int y = 0)
     : xPos(x), yPos(y) {}
     
-    float calculateMagnitude(){
+    int calculateMagnitude(){
         return sqrt(pow(xPos,2) + pow(yPos,2));
     }
     
-    float getAngle(){
+    int getAngle(){
         return atan2(yPos, xPos) * 180/M_PI;
     }
     
     // Calculate the difference vector and return its magnitude
     // Used for target's distance walked 
-    float getDifferenceVectorMag(const vector vec){
+    int getDifferenceVectorMag(const vector vec){
         vector aux = vector((xPos - vec.xPos), (yPos - vec.yPos));
         return aux.calculateMagnitude();
     }
@@ -29,8 +29,8 @@ struct vector{
     }
     
     bool operator>(vector &vec){
-        float magnitude = calculateMagnitude();
-        float vecMagnitude = vec.calculateMagnitude();
+        int magnitude = calculateMagnitude();
+        int vecMagnitude = vec.calculateMagnitude();
         
         return (magnitude > vecMagnitude);
     }
@@ -39,24 +39,24 @@ struct vector{
         return !(*this > vec);
     }
     
-    float operator+(vector &vec){
+    int operator+(vector &vec){
         return getAngle() - vec.getAngle();
     }
     
     // Angular error
-    float operator-(vector &vec){
+    int operator-(vector &vec){
         return getAngle() - vec.getAngle();
     }
     
-    operator double const(){
+    operator int const(){
         return getAngle();
     }
 };
 
 // Calculate velocity of the target to imitate it
-float calculateVelocity(vector oldSample, vector newSample, long &lastTime, int DEBUG = 0){
-    float distanceRun = newSample.getDifferenceVectorMag(oldSample);
-    float deltaTime = (millis() - lastTime) / 1000.0; // Seconds
+int calculateVelocity(vector oldSample, vector newSample, long &lastTime, int DEBUG = 0){
+    int distanceRun = newSample.getDifferenceVectorMag(oldSample);
+    int deltaTime = (millis() - lastTime) / 1000; // Seconds
     int velocity =  distanceRun / deltaTime; // cm/s
 
     if (DEBUG){
@@ -69,13 +69,13 @@ float calculateVelocity(vector oldSample, vector newSample, long &lastTime, int 
     return velocity;
 }
 
-float calculateVelocity(vector newSample){
-  float distance = newSample.calculateMagnitude();
+int calculateVelocity(vector newSample){
+  int distance = newSample.calculateMagnitude();
   int minDistance = 20;
   int maxDistance = 50;
   int minPWM = 0;
   int maxPWM = 200;
-  float velocity = map(distance, minDistance, maxDistance, minPWM, maxPWM);
+  int velocity = map(distance, minDistance, maxDistance, minPWM, maxPWM);
 
   return velocity;
 }
